Moved the get_example_xyz() function to FOX.functions.utils
