+++ FOX/functions/__init__.py
@@ -1,13 +1,13 @@
 from .rdf import (get_rdf_lowmem, get_rdf)
 from .adf import get_adf
 from .read_xyz import read_multi_xyz
-from .utils import (get_template, assert_error)
+from .utils import (get_template, assert_error, get_example_xyz)
 from .hdf5_utils import (create_hdf5, to_hdf5, from_hdf5)
 
 __all__ = [
     'get_rdf_lowmem', 'get_rdf',
     'get_adf',
     'read_multi_xyz',
-    'get_template', 'assert_error',
+    'get_template', 'assert_error', 'get_example_xyz',
     'create_hdf5', 'to_hdf5', 'from_hdf5'
 ]
